Standardize Arroyo's use of DataFusion's extensions on ArroyoExtension.
