+++ crates/arroyo-df/src/extension/watermark_node.rs
@@ -1,10 +1,18 @@
+use crate::builder::{NamedNode, Planner};
+use crate::extension::{ArroyoExtension, NodeWithIncomingEdges};
 use crate::schemas::add_timestamp_field;
-use anyhow::{anyhow, Result};
-use arroyo_rpc::df::ArroyoSchema;
+use anyhow::anyhow;
+use arroyo_datastream::logical::{LogicalEdge, LogicalEdgeType, LogicalNode, OperatorName};
+use arroyo_rpc::df::{ArroyoSchema, ArroyoSchemaRef};
+use arroyo_rpc::grpc::api::ExpressionWatermarkConfig;
 use datafusion_common::{DFSchemaRef, OwnedTableReference};
 use datafusion_expr::{Expr, LogicalPlan, UserDefinedLogicalNodeCore};
-use std::{fmt::Formatter, sync::Arc};
+use datafusion_proto::protobuf::PhysicalExprNode;
+use prost::Message;
+use std::fmt::Formatter;
+use std::sync::Arc;
 
+pub(crate) const WATERMARK_NODE_NAME: &'static str = "WatermarkNode";
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct WatermarkNode {
     pub input: LogicalPlan,
@@ -16,7 +24,7 @@ pub struct WatermarkNode {
 
 impl UserDefinedLogicalNodeCore for WatermarkNode {
     fn name(&self) -> &str {
-        "WatermarkNode"
+        WATERMARK_NODE_NAME
     }
 
     fn inputs(&self) -> Vec<&LogicalPlan> {
@@ -64,12 +72,50 @@ impl UserDefinedLogicalNodeCore for WatermarkNode {
     }
 }
 
+impl ArroyoExtension for WatermarkNode {
+    fn node_name(&self) -> Option<NamedNode> {
+        Some(NamedNode::Watermark(self.qualifier.clone()))
+    }
+
+    fn plan_node(
+        &self,
+        planner: &Planner,
+        index: usize,
+        input_schemas: Vec<ArroyoSchemaRef>,
+    ) -> anyhow::Result<NodeWithIncomingEdges> {
+        let expression = planner.create_physical_expr(&self.watermark_expression, &self.schema)?;
+        let expression = PhysicalExprNode::try_from(expression)?;
+        let node = LogicalNode {
+            operator_id: format!("watermark_{}", index),
+            description: "watermark".to_string(),
+            operator_name: OperatorName::ExpressionWatermark,
+            parallelism: 1,
+            operator_config: ExpressionWatermarkConfig {
+                period_micros: 1_000_000,
+                idle_time_micros: None,
+                expression: expression.encode_to_vec(),
+                input_schema: Some(self.arroyo_schema().try_into().unwrap()),
+            }
+            .encode_to_vec(),
+        };
+        let incoming_edge =
+            LogicalEdge::project_all(LogicalEdgeType::Forward, input_schemas[0].as_ref().clone());
+        Ok(NodeWithIncomingEdges {
+            node,
+            edges: vec![incoming_edge],
+        })
+    }
+    fn output_schema(&self) -> ArroyoSchema {
+        self.arroyo_schema()
+    }
+}
+
 impl WatermarkNode {
     pub(crate) fn new(
         input: LogicalPlan,
         qualifier: OwnedTableReference,
         watermark_expression: Expr,
-    ) -> Result<Self> {
+    ) -> anyhow::Result<Self> {
         let schema = add_timestamp_field(input.schema().clone(), Some(qualifier.clone()))?;
         let timestamp_index = schema
             .index_of_column_by_name(None, "_timestamp")?
