+++ crates/arroyo-df/src/plan.rs
@@ -1,446 +1,31 @@
-use std::{collections::HashMap, fmt::Formatter, sync::Arc};
+use std::sync::Arc;
 
 use anyhow::Result;
-use arrow::datatypes::IntervalMonthDayNanoType;
 
 use arroyo_datastream::WindowType;
 use arroyo_rpc::TIMESTAMP_FIELD;
 use datafusion_common::{
     plan_err,
     tree_node::{TreeNode, TreeNodeRewriter, TreeNodeVisitor, VisitRecursion},
-    Column, DFField, DFSchema, DFSchemaRef, DataFusionError, OwnedTableReference,
-    Result as DFResult, ScalarValue,
+    Column, DFField, DFSchema, DataFusionError, Result as DFResult, ScalarValue,
 };
-use datafusion_execution::FunctionRegistry;
+
 use datafusion_expr::{
     expr::{Alias, ScalarFunction},
-    Aggregate, BinaryExpr, BuiltinScalarFunction, Case, DmlStatement, Expr, Extension, Join,
-    JoinConstraint, JoinType, LogicalPlan, Projection, ScalarFunctionDefinition,
-    UserDefinedLogicalNodeCore, WriteOp,
+    Aggregate, BinaryExpr, BuiltinScalarFunction, Case, Expr, Extension, Join, JoinConstraint,
+    JoinType, LogicalPlan, Projection,
 };
 
 use crate::{
+    extension::{
+        aggregate::AggregateExtension, join::JoinExtension,
+        key_calculation::KeyCalculationExtension,
+    },
     find_window,
-    physical::new_registry,
     rewriters::SourceRewriter,
     schemas::{add_timestamp_field, has_timestamp_field},
-    tables::{ConnectorTable, Table},
     ArroyoSchemaProvider, WindowBehavior,
 };
-
-/* Calculation for computing keyed data, with a vec of keys
-   that will be used for shuffling data to the correct nodes.
-*/
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) struct KeyCalculationExtension {
-    pub(crate) name: Option<String>,
-    pub(crate) input: LogicalPlan,
-    pub(crate) keys: Vec<usize>,
-    pub(crate) schema: DFSchemaRef,
-}
-
-impl KeyCalculationExtension {
-    pub fn new_named_and_trimmed(input: LogicalPlan, keys: Vec<usize>, name: String) -> Self {
-        let output_fields: Vec<_> = input
-            .schema()
-            .fields()
-            .iter()
-            .enumerate()
-            .filter_map(|(index, field)| {
-                if !keys.contains(&index) {
-                    Some(field.clone())
-                } else {
-                    None
-                }
-            })
-            .collect();
-        let schema =
-            DFSchema::new_with_metadata(output_fields, input.schema().metadata().clone()).unwrap();
-        Self {
-            name: Some(name),
-            input,
-            keys,
-            schema: Arc::new(schema),
-        }
-    }
-    pub fn new(input: LogicalPlan, keys: Vec<usize>) -> Self {
-        let schema = input.schema().clone();
-        Self {
-            name: None,
-            input,
-            keys,
-            schema,
-        }
-    }
-}
-
-impl UserDefinedLogicalNodeCore for KeyCalculationExtension {
-    fn name(&self) -> &str {
-        "KeyCalculationExtension"
-    }
-
-    fn inputs(&self) -> Vec<&LogicalPlan> {
-        vec![&self.input]
-    }
-
-    fn schema(&self) -> &DFSchemaRef {
-        &self.schema
-    }
-
-    fn expressions(&self) -> Vec<Expr> {
-        vec![]
-    }
-
-    fn fmt_for_explain(&self, f: &mut Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "KeyCalculationExtension: {}",
-            self.schema()
-                .fields()
-                .iter()
-                .map(|f| f.qualified_name())
-                .collect::<Vec<_>>()
-                .join(", ")
-        )
-    }
-
-    fn from_template(&self, _exprs: &[Expr], inputs: &[LogicalPlan]) -> Self {
-        assert_eq!(inputs.len(), 1, "input size inconsistent");
-        match self.name {
-            Some(ref name) => {
-                Self::new_named_and_trimmed(inputs[0].clone(), self.keys.clone(), name.clone())
-            }
-            None => Self::new(inputs[0].clone(), self.keys.clone()),
-        }
-    }
-}
-
-/* Lightweight extension that allows us to segment the graph and merge nodes with the same name.
-  An Extension Planner will be used to isolate computation to individual nodes.
-*/
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) struct RemoteTableExtension {
-    pub(crate) input: LogicalPlan,
-    pub(crate) name: OwnedTableReference,
-    pub(crate) schema: DFSchemaRef,
-}
-
-impl UserDefinedLogicalNodeCore for RemoteTableExtension {
-    fn name(&self) -> &str {
-        "RemoteTableExtension"
-    }
-
-    fn inputs(&self) -> Vec<&LogicalPlan> {
-        vec![&self.input]
-    }
-
-    fn schema(&self) -> &DFSchemaRef {
-        &self.schema
-    }
-
-    fn expressions(&self) -> Vec<Expr> {
-        vec![]
-    }
-
-    fn fmt_for_explain(&self, f: &mut Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "RemoteTableExtension: {}",
-            self.schema
-                .fields()
-                .iter()
-                .map(|f| f.qualified_name())
-                .collect::<Vec<_>>()
-                .join(", ")
-        )
-    }
-
-    fn from_template(&self, _exprs: &[Expr], inputs: &[LogicalPlan]) -> Self {
-        Self {
-            input: inputs[0].clone(),
-            name: self.name.clone(),
-            schema: self.schema.clone(),
-        }
-    }
-}
-
-/*
-This is a plan used for appending a _timestamp field to an existing record batch.
- */
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) struct TimestampAppendExtension {
-    pub(crate) input: LogicalPlan,
-    pub(crate) qualifier: Option<OwnedTableReference>,
-    pub(crate) schema: DFSchemaRef,
-}
-
-impl TimestampAppendExtension {
-    fn new(input: LogicalPlan, qualifier: Option<OwnedTableReference>) -> Self {
-        if has_timestamp_field(input.schema().clone()) {
-            unreachable!("shouldn't be adding timestamp to a plan that already has it");
-        }
-        let schema = add_timestamp_field(input.schema().clone(), qualifier.clone()).unwrap();
-        Self {
-            input,
-            qualifier,
-            schema,
-        }
-    }
-}
-
-impl UserDefinedLogicalNodeCore for TimestampAppendExtension {
-    fn name(&self) -> &str {
-        "TimestampAppendExtension"
-    }
-
-    fn inputs(&self) -> Vec<&LogicalPlan> {
-        vec![&self.input]
-    }
-
-    fn schema(&self) -> &DFSchemaRef {
-        &self.schema
-    }
-
-    fn expressions(&self) -> Vec<Expr> {
-        vec![]
-    }
-
-    fn fmt_for_explain(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "TimestampAppendExtension({:?}): {}",
-            self.qualifier,
-            self.schema
-                .fields()
-                .iter()
-                .map(|f| f.qualified_name())
-                .collect::<Vec<_>>()
-                .join(", ")
-        )
-    }
-
-    fn from_template(&self, _exprs: &[Expr], inputs: &[LogicalPlan]) -> Self {
-        Self::new(inputs[0].clone(), self.qualifier.clone())
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct WindowAppendExtension {
-    pub(crate) input: LogicalPlan,
-    pub(crate) window_field: DFField,
-    pub(crate) window_index: usize,
-    pub(crate) schema: DFSchemaRef,
-}
-
-impl WindowAppendExtension {
-    fn new(input: LogicalPlan, window_field: DFField, window_index: usize) -> Self {
-        let mut fields = input.schema().fields().clone();
-        fields.insert(window_index, window_field.clone());
-        let metadata = input.schema().metadata().clone();
-        Self {
-            input,
-            window_field,
-            window_index,
-            schema: Arc::new(DFSchema::new_with_metadata(fields, metadata).unwrap()),
-        }
-    }
-}
-
-impl UserDefinedLogicalNodeCore for WindowAppendExtension {
-    fn name(&self) -> &str {
-        "WindowAppendExtension"
-    }
-
-    fn inputs(&self) -> Vec<&LogicalPlan> {
-        vec![&self.input]
-    }
-
-    fn schema(&self) -> &DFSchemaRef {
-        &self.schema
-    }
-
-    fn expressions(&self) -> Vec<Expr> {
-        vec![]
-    }
-
-    fn fmt_for_explain(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "WindowAppendExtension: field {:?} at {}",
-            self.window_field, self.window_index
-        )
-    }
-
-    fn from_template(&self, _exprs: &[Expr], inputs: &[LogicalPlan]) -> Self {
-        Self::new(
-            inputs[0].clone(),
-            self.window_field.clone(),
-            self.window_index,
-        )
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) struct AggregateExtension {
-    pub(crate) window_behavior: WindowBehavior,
-    pub(crate) aggregate: LogicalPlan,
-    pub(crate) schema: DFSchemaRef,
-    pub(crate) key_fields: Vec<usize>,
-    pub(crate) final_calculation: LogicalPlan,
-}
-
-impl AggregateExtension {
-    pub fn new(
-        window_behavior: WindowBehavior,
-        aggregate: LogicalPlan,
-        key_fields: Vec<usize>,
-    ) -> Self {
-        let final_calculation =
-            Self::final_projection(&aggregate, window_behavior.clone()).unwrap();
-
-        Self {
-            window_behavior,
-            aggregate,
-            schema: final_calculation.schema().clone(),
-            key_fields,
-            final_calculation,
-        }
-    }
-
-    // projection assuming that _timestamp has been populated with the start of the bin.
-    pub fn final_projection(
-        aggregate_plan: &LogicalPlan,
-        window_behavior: WindowBehavior,
-    ) -> DFResult<LogicalPlan> {
-        let timestamp_field = aggregate_plan.inputs()[0]
-            .schema()
-            .field_with_unqualified_name(TIMESTAMP_FIELD)
-            .unwrap()
-            .clone();
-        let timestamp_append = LogicalPlan::Extension(Extension {
-            node: Arc::new(TimestampAppendExtension::new(
-                aggregate_plan.clone(),
-                timestamp_field.qualifier().cloned(),
-            )),
-        });
-        let mut aggregate_fields = aggregate_plan.schema().fields().clone();
-        let mut aggregate_expressions: Vec<_> = aggregate_fields
-            .iter()
-            .map(|field| Expr::Column(field.qualified_column()))
-            .collect();
-        let (window_field, window_index, width) = match window_behavior {
-            WindowBehavior::InData => return Ok(timestamp_append),
-            WindowBehavior::FromOperator {
-                window,
-                window_field,
-                window_index,
-            } => match window {
-                WindowType::Tumbling { width, .. } | WindowType::Sliding { width, .. } => {
-                    (window_field, window_index, width)
-                }
-                WindowType::Session { .. } => {
-                    return Ok(LogicalPlan::Extension(Extension {
-                        node: Arc::new(WindowAppendExtension::new(
-                            timestamp_append,
-                            window_field,
-                            window_index,
-                        )),
-                    }))
-                }
-                WindowType::Instant => return Ok(timestamp_append),
-            },
-        };
-
-        let timestamp_column =
-            Column::new(timestamp_field.qualifier().cloned(), timestamp_field.name());
-        aggregate_fields.insert(window_index, window_field.clone());
-        let registry = new_registry();
-        let window_expression = Expr::ScalarFunction(ScalarFunction {
-            func_def: ScalarFunctionDefinition::UDF(registry.udf("window")?),
-            args: vec![
-                // copy bin_start as first argument
-                Expr::Column(timestamp_column.clone()),
-                // add width interval to _timestamp for bin end
-                Expr::BinaryExpr(BinaryExpr {
-                    left: Box::new(Expr::Column(timestamp_column.clone())),
-                    op: datafusion_expr::Operator::Plus,
-                    right: Box::new(Expr::Literal(ScalarValue::IntervalMonthDayNano(Some(
-                        IntervalMonthDayNanoType::make_value(0, 0, width.as_nanos() as i64),
-                    )))),
-                }),
-            ],
-        });
-        aggregate_expressions.insert(
-            window_index,
-            window_expression
-                .alias_qualified(window_field.qualifier().cloned(), window_field.name()),
-        );
-        aggregate_fields.push(timestamp_field.clone());
-        let bin_end_calculation = Expr::BinaryExpr(BinaryExpr {
-            left: Box::new(Expr::Column(timestamp_column.clone())),
-            op: datafusion_expr::Operator::Plus,
-            right: Box::new(Expr::Literal(ScalarValue::IntervalMonthDayNano(Some(
-                IntervalMonthDayNanoType::make_value(0, 0, (width.as_nanos() - 1) as i64),
-            )))),
-        });
-        aggregate_expressions.push(bin_end_calculation);
-        Ok(LogicalPlan::Projection(
-            datafusion_expr::Projection::try_new_with_schema(
-                aggregate_expressions,
-                Arc::new(timestamp_append),
-                Arc::new(DFSchema::new_with_metadata(
-                    aggregate_fields,
-                    HashMap::new(),
-                )?),
-            )?,
-        ))
-    }
-}
-
-impl UserDefinedLogicalNodeCore for AggregateExtension {
-    fn name(&self) -> &str {
-        "AggregateExtension"
-    }
-
-    fn inputs(&self) -> Vec<&LogicalPlan> {
-        vec![&self.aggregate]
-    }
-
-    fn schema(&self) -> &DFSchemaRef {
-        &self.schema
-    }
-
-    fn expressions(&self) -> Vec<Expr> {
-        vec![]
-    }
-
-    fn fmt_for_explain(&self, f: &mut Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "AggregateExtension: {} | window_behavior: {:?}",
-            self.schema()
-                .fields()
-                .iter()
-                .map(|f| f.qualified_name())
-                .collect::<Vec<_>>()
-                .join(", "),
-            match &self.window_behavior {
-                WindowBehavior::InData => "InData".to_string(),
-                WindowBehavior::FromOperator { window, .. } =>
-                    format!("FromOperator({:?})", window),
-            }
-        )
-    }
-
-    fn from_template(&self, _exprs: &[Expr], inputs: &[LogicalPlan]) -> Self {
-        assert_eq!(inputs.len(), 1, "input size inconsistent");
-        Self::new(
-            self.window_behavior.clone(),
-            inputs[0].clone(),
-            self.key_fields.clone(),
-        )
-    }
-}
 
 #[derive(Debug, Default)]
 pub struct AggregateRewriter {}
@@ -652,211 +237,6 @@ impl TreeNodeVisitor for WindowDetectingVisitor {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) struct TableSourceExtension {
-    pub(crate) name: OwnedTableReference,
-    pub(crate) table: ConnectorTable,
-    pub(crate) schema: DFSchemaRef,
-}
-
-impl TableSourceExtension {
-    pub fn new(name: OwnedTableReference, table: ConnectorTable) -> Self {
-        let physical_fields = table
-            .fields
-            .iter()
-            .filter_map(|field| match field {
-                crate::tables::FieldSpec::StructField(field) => {
-                    Some(DFField::from_qualified(&name, Arc::new(field.clone())))
-                }
-                crate::tables::FieldSpec::VirtualField { .. } => None,
-            })
-            .collect::<Vec<_>>();
-        let schema =
-            Arc::new(DFSchema::new_with_metadata(physical_fields, HashMap::new()).unwrap());
-        let schema = add_timestamp_field(schema, Some(name.clone())).unwrap();
-        Self {
-            name,
-            table,
-            schema,
-        }
-    }
-}
-
-impl UserDefinedLogicalNodeCore for TableSourceExtension {
-    fn name(&self) -> &str {
-        "TableSourceExtension"
-    }
-
-    fn inputs(&self) -> Vec<&LogicalPlan> {
-        vec![]
-    }
-
-    fn schema(&self) -> &DFSchemaRef {
-        &self.schema
-    }
-
-    fn expressions(&self) -> Vec<Expr> {
-        vec![]
-    }
-
-    fn fmt_for_explain(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "TableSourceExtension: {}",
-            self.schema
-                .fields()
-                .iter()
-                .map(|f| f.qualified_name())
-                .collect::<Vec<_>>()
-                .join(", ")
-        )
-    }
-
-    fn from_template(&self, _exprs: &[Expr], _inputs: &[LogicalPlan]) -> Self {
-        Self {
-            name: self.name.clone(),
-            table: self.table.clone(),
-            schema: self.schema.clone(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) struct SinkExtension {
-    pub(crate) name: OwnedTableReference,
-    pub(crate) table: Table,
-    pub(crate) schema: DFSchemaRef,
-    pub(crate) input: Arc<LogicalPlan>,
-}
-
-impl UserDefinedLogicalNodeCore for SinkExtension {
-    fn name(&self) -> &str {
-        "SinkExtension"
-    }
-
-    fn inputs(&self) -> Vec<&LogicalPlan> {
-        vec![&self.input]
-    }
-
-    fn schema(&self) -> &DFSchemaRef {
-        &self.schema
-    }
-
-    fn expressions(&self) -> Vec<Expr> {
-        vec![]
-    }
-
-    fn fmt_for_explain(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "SinkExtension({:?}): {}",
-            self.name,
-            self.schema
-                .fields()
-                .iter()
-                .map(|f| f.qualified_name())
-                .collect::<Vec<_>>()
-                .join(", ")
-        )
-    }
-
-    fn from_template(&self, _exprs: &[Expr], inputs: &[LogicalPlan]) -> Self {
-        assert_eq!(inputs.len(), 1, "input size inconsistent");
-        Self {
-            name: self.name.clone(),
-            table: self.table.clone(),
-            schema: self.schema.clone(),
-            input: Arc::new(inputs[0].clone()),
-        }
-    }
-}
-
-/// Rewrites a logical plan to replace InsertIntos with a SinkExtension that knows everything about Arroyo sinks.
-pub struct SinkRewriter {
-    pub(crate) schema_provider: ArroyoSchemaProvider,
-}
-
-impl TreeNodeRewriter for SinkRewriter {
-    type N = LogicalPlan;
-
-    fn mutate(&mut self, node: Self::N) -> DFResult<Self::N> {
-        match &node {
-            LogicalPlan::Dml(DmlStatement {
-                table_name,
-                table_schema,
-                op,
-                input,
-            }) => {
-                let WriteOp::InsertInto = op else {
-                    // error out
-                    return Err(DataFusionError::Plan(
-                        "only InsertInto is supported".to_string(),
-                    ));
-                };
-                let table = self
-                    .schema_provider
-                    .get_table(table_name.to_string())
-                    .ok_or_else(|| {
-                        DataFusionError::Plan(format!("table {} not found", table_name))
-                    })?;
-                let sink_extension = SinkExtension {
-                    name: table_name.clone(),
-                    table: table.clone(),
-                    schema: table_schema.clone(),
-                    input: input.clone(),
-                };
-                Ok(LogicalPlan::Extension(Extension {
-                    node: Arc::new(sink_extension),
-                }))
-            }
-            _ => Ok(node),
-        }
-    }
-}
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) struct JoinExtension {
-    pub(crate) rewritten_join: LogicalPlan,
-    pub(crate) is_instant: bool,
-}
-
-impl UserDefinedLogicalNodeCore for JoinExtension {
-    fn name(&self) -> &str {
-        "JoinExtension"
-    }
-
-    fn inputs(&self) -> Vec<&LogicalPlan> {
-        vec![&self.rewritten_join]
-    }
-
-    fn schema(&self) -> &DFSchemaRef {
-        self.rewritten_join.schema()
-    }
-
-    fn expressions(&self) -> Vec<Expr> {
-        vec![]
-    }
-
-    fn fmt_for_explain(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "JoinExtension: {}",
-            self.schema()
-                .fields()
-                .iter()
-                .map(|f| f.qualified_name())
-                .collect::<Vec<_>>()
-                .join(", ")
-        )
-    }
-
-    fn from_template(&self, _exprs: &[Expr], inputs: &[LogicalPlan]) -> Self {
-        Self {
-            rewritten_join: inputs[0].clone(),
-            is_instant: self.is_instant,
-        }
-    }
-}
-
 pub(crate) struct JoinRewriter {}
 
 impl JoinRewriter {
@@ -918,7 +298,7 @@ impl JoinRewriter {
         let fields = projection
             .schema
             .fields()
-            .into_iter()
+            .iter()
             .enumerate()
             .map(|(index, field)| {
                 // rename to avoid collisions
@@ -940,7 +320,7 @@ impl JoinRewriter {
         projection.schema = rewritten_schema;
         let key_calculation_extension = KeyCalculationExtension::new_named_and_trimmed(
             LogicalPlan::Projection(projection),
-            (0..key_count).into_iter().collect(),
+            (0..key_count).collect(),
             name.to_string(),
         );
         Ok(LogicalPlan::Extension(Extension {
@@ -1113,7 +493,7 @@ impl<'a> TreeNodeRewriter for ArroyoRewriter<'a> {
             }
             LogicalPlan::TableScan(table_scan) => {
                 return SourceRewriter {
-                    schema_provider: &self.schema_provider,
+                    schema_provider: self.schema_provider,
                 }
                 .mutate(LogicalPlan::TableScan(table_scan));
             }
